Issue #130: Properties with dashes in their names
Everything appears to work as it did previously, but this modification now handles the use case where the declarations variable could have a dash in it (EX: "value:background-color"). Additionally the declarations variable is passed into the function, to ensure it's defined.

This pull request includes the type conversion that happened automatically by creating a function using the function constructor.